fix: remove environment-edit-view folder.
